JAMP 32-bit linux uses i386 for filename in openjamp (but not original jamp server), so we'll go with that. also remove the typecast lambdas for return values. the original mod caller will get the return values from the proper locations without us having to do it (plus orientation_t return type messes up the call stack). also add the actual int cmd value to all the syscall/vmmain TRACE log lines

+++ src/plugin.cpp
@@ -367,7 +367,7 @@ static int* s_plugin_helper_ConfigGetArrayInt(const char* key) {
 	index = (index + 1) & 7;
 	value[index] = cfg_get_array_int(node, key);
 	// insert length of the array as the first element
-	value[index].insert(value[index].begin(), value[index].size());
+	value[index].insert(value[index].begin(), (int)value[index].size());
 	return value[index].data();
 }
